Consolidate implementing testing files
